XML files loaded in
XML files now load in at start of application. Will allow user adjustment of available automata between runs. Added a couple others into the list available also for the hell of it.

I've left Conway's game of life and Langtons ant hard coded. The former just because it felt a good example for the hardcoded variant, the latter for speed optimization.